requirements.txt
appdirs>=1.4.3
asgiref>=3.3.1
Babel>=2.9.0
CacheControl>=0.12.6
certifi>=2019.11.28
chardet>=3.0.4
colorama>=0.4.3
contextlib2>=0.6.0
coreapi>=2.3.3
coreschema>=0.0.4
defusedxml>=0.7.1
diff-match-patch>=20200713
distlib>=0.3.0
distro>=1.4.0
Django>=3.1.7
django-ckeditor>=6.0.0
django-cors-headers>=3.7.0
django-faicon>=0.1.0
django-filter>=2.4.0
django-import-export>=2.5.0
django-js-asset>=1.2.2
django-phonenumber-field>=5.0.0
djangorestframework>=3.12.2
drf-yasg>=1.20.0
et-xmlfile>=1.0.1
gunicorn>=20.0.4
html5lib>=1.0.1
idna>=2.8
inflection>=0.5.1
ipaddr>=2.2.0
itypes>=1.2.0
Jinja2>=2.11.3
lockfile>=0.12.2
Markdown>=3.3.4
MarkupPy>=1.14
MarkupSafe>=1.1.1
msgpack>=0.6.2
odfpy>=1.4.1
openapi-codec>=1.3.2
openpyxl>=3.0.7
packaging>=20.3
pep517>=0.8.2
phonenumbers>=8.12.19
Pillow>=8.1.1
# pkg-resources>=0.0.0
progress>=1.5
psycopg2-binary>=2.8.6
pyparsing>=2.4.6
pytoml>=0.1.21
pytz>=2021.1
PyYAML>=5.4.1
requests>=2.22.0
retrying>=1.3.3
ruamel.yaml>=0.16.13
ruamel.yaml.clib>=0.2.2
simplejson>=3.17.2
six>=1.14.0
sqlparse>=0.4.1
tablib>=3.0.0
Unidecode>=1.2.0
uritemplate>=3.0.1
urllib3>=1.25.8
webencodings>=0.5.1
xlrd>=2.0.1
xlwt>=1.3.0
